+++ packages/schema-to-validator/package.json
@@ -1,15 +1,17 @@
-export default {
-  "name": "@traversable/derive-validators",
+{
+  "name": "@traversable/schema-to-validator",
   "type": "module",
-  "version": "0.0.31",
+  "version": "0.0.0",
   "private": false,
   "description": "",
   "license": "MIT",
-  "sideEffects": ["./src/exports.ts"],
+  "sideEffects": [
+    "./src/exports.ts"
+  ],
   "repository": {
     "type": "git",
     "url": "https://github.com/traversable/schema.git",
-    "directory": "packages/derive-validators"
+    "directory": "packages/schema-to-validator"
   },
   "bugs": {
     "url": "https://github.com/traversable/schema/issues",
@@ -17,10 +19,14 @@ export default {
   },
   "@traversable": {
     "generateExports": {
-      "include": ["**/*.ts"]
+      "include": [
+        "**/*.ts"
+      ]
     },
     "generateIndex": {
-      "include": ["**/*.ts"]
+      "include": [
+        "**/*.ts"
+      ]
     }
   },
   "publishConfig": {
@@ -51,4 +57,4 @@ export default {
     "@traversable/schema": "workspace:^",
     "@traversable/schema-seed": "workspace:^"
   }
-} as const
+}
