+++ packages/schema-to-validator/src/errors.ts
