+++ packages/schema-to-validator/src/shared.ts
