+++ packages/schema-to-validator/tsconfig.build.json
