+++ packages/schema-to-validator/tsconfig.json
