+++ packages/schema-to-validator/tsconfig.src.json
