+++ packages/schema-to-validator/tsconfig.test.json
