+++ packages/schema-to-validator/vite.config.ts
