+++ tsconfig.build.json
@@ -5,7 +5,6 @@
     { "path": "packages/arktype-test/tsconfig.build.json" },
     { "path": "packages/arktype-types/tsconfig.build.json" },
     { "path": "packages/arktype/tsconfig.build.json" },
-    { "path": "packages/derive-validators/tsconfig.build.json" },
     { "path": "packages/json-schema-test/tsconfig.build.json" },
     { "path": "packages/json-schema-types/tsconfig.build.json" },
     { "path": "packages/json-schema/tsconfig.build.json" },
@@ -18,6 +17,7 @@
     { "path": "packages/schema-seed/tsconfig.build.json" },
     { "path": "packages/schema-to-json-schema/tsconfig.build.json" },
     { "path": "packages/schema-to-string/tsconfig.build.json" },
+    { "path": "packages/schema-to-validator/tsconfig.build.json" },
     { "path": "packages/schema-valibot-adapter/tsconfig.build.json" },
     { "path": "packages/schema/tsconfig.build.json" },
     { "path": "packages/typebox-test/tsconfig.build.json" },
