+++ tsconfig.json
@@ -5,7 +5,6 @@
     { "path": "packages/arktype" },
     { "path": "packages/arktype-test" },
     { "path": "packages/arktype-types" },
-    { "path": "packages/derive-validators" },
     { "path": "packages/json" },
     { "path": "packages/json-schema" },
     { "path": "packages/json-schema-test" },
@@ -19,6 +18,7 @@
     { "path": "packages/schema-seed" },
     { "path": "packages/schema-to-json-schema" },
     { "path": "packages/schema-to-string" },
+    { "path": "packages/schema-to-validator" },
     { "path": "packages/schema-valibot-adapter" },
     { "path": "packages/typebox" },
     { "path": "packages/typebox-test" },
